# Correspondence_backend/src/swagger/swagger.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "Correspondence System API Documentation",
      "description": "Fake Online REST API for Testing and Prototyping",
      "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:5000/api/",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/inbox/paging/v2" : {
            "get" : {
                "summary" : "Get all threads",
                "description": "Get all threads based on the value of the page and limit parameter",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "query",
                        "name": "page",
                        "description": "page parameter represents the current page location",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    { 
                        "in": "query",
                        "name": "limit",
                        "description": "page parameter represents the current page location",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Got all inbox threads successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/correspondence"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    }
                }
            }
        },
        "/inbox/count/v2": {
            "get": {
                "summary": "Get the count of all threads",
                "description": "Get the count of all threads",
                "produces": ["application/json"],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "count": {
                                    "type": "integer"
                                }
                            }
                        }
                    }, 
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    },
                    "500": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "correspondence": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "corr_no": {
                    "type": "string"
                },
                "corr_type": {
                    "type": "string"
                },
                "entity_no": {
                    "type": "string"
                },
                "from_user": {
                    "type": "array"
                },
                "cc_entity": {
                    "type": "array"
                },
                "to_entity": {
                    "type": "string"
                },
                "to_department": {
                    "type": "string"
                },
                "sent_date": {
                    "type": "string"
                },
                "corr_subject": {
                    "type": "string"
                },
                "corr_body": {
                    "type": "string"
                },
                "message_status": {
                    "type": "string"
                },
                "docs_IDs": {
                    "type": "array"
                },
                "due_date": {
                    "type": "string"
                },
                "delete": {
                    "type": "boolean"
                },
                "draft": {
                    "type": "boolean"
                },
                "thread_id": {
                    "type": "string"
                },
                "await_reply": {
                    "type": "boolean"
                },
                "classification": {
                    "type": "string"
                },
                "priority": {
                    "type": "string"
                },
                "received_date": {
                    "type": "string"
                },
                "starred": {
                    "type": "boolean"
                },
                "isThreadStarred": {
                    "type": "boolean"
                },
                "isThreadDrafted": {
                    "type": "boolean"
                }
            }
        }
    }
  }